reorganize enchantments
