+++ src/main/java/gregtech/api/unification/crafttweaker/MaterialExpansion.java
@@ -6,7 +6,6 @@ import crafttweaker.api.enchantments.IEnchantment;
 import crafttweaker.api.liquid.ILiquidDefinition;
 import crafttweaker.api.minecraft.CraftTweakerMC;
 import gregtech.api.GTValues;
-import gregtech.api.enchants.EnchantmentData;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.material.info.MaterialFlag;
 import gregtech.api.unification.material.info.MaterialIconSet;
@@ -192,7 +191,7 @@ public class MaterialExpansion {
         ToolProperty prop = m.getProperty(PropertyKey.TOOL);
         if (prop != null) {
             Enchantment enchantmentType = (Enchantment) enchantment.getDefinition().getInternal();
-            prop.toolEnchantments.add(new EnchantmentData(enchantmentType, enchantment.getLevel()));
+            prop.addEnchantmentForTools(enchantmentType, enchantment.getLevel());
         } else logError(m, "change tool enchantments", "Tool");
     }
