+++ src/main/java/gregtech/modules/GregTechModules.java
@@ -5,6 +5,7 @@ import gregtech.api.modules.IModuleContainer;
 public class GregTechModules implements IModuleContainer {
 
     public static final String MODULE_CORE = "core";
+    public static final String MODULE_TOOLS = "tools";
 
     // Integration modules
     public static final String MODULE_JEI = "jei_integration";
